use UInt64 for totalCapacity in storagepool

+++ templates/storagepool/schema.capnp
@@ -2,7 +2,7 @@
 
 struct Schema {
     status @0 :Status;
-    totalCapacity @1:UInt32;
+    totalCapacity @1:UInt64;
     freeCapacity @2:UInt32;
     metadataProfile @3 :Profile;
     dataProfile @4 :Profile;
